+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Use `withOwnerId` directive on `searchOrderForm` query.
+
 ## [2.161.1] - 2023-01-24
 
 ## [2.161.0] - 2023-01-17